+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/referee_game_files.js
@@ -136,6 +136,9 @@ define([
         }
         return attribute;
       },
+      multiplyOrAdd: function (attribute, value) {
+        return _.isNumber(attribute) ? attribute * value : value;
+      },
     };
 
     const applyMod = function (mod) {