+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -16,7 +16,8 @@ define([
     "gwaio_enable_suddendeath",
     "gwaio_enable_landanywhere",
     "gwaio_enable_orbitalbombardment",
-    "gwaio_anti_tank"
+    "gwaio_anti_tank",
+    "gwaio_anti_bots"
   );
 
   return {
@@ -450,6 +451,7 @@ define([
       { id: "gwaio_enable_landanywhere" },
       { id: "gwaio_enable_orbitalbombardment" },
       { id: "gwaio_anti_tank" },
+      { id: "gwaio_anti_bots" },
     ],
   };
 });