+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -226,6 +226,7 @@ function gwoCard() {
       const expandedCards = [
         "gwaio_anti_bots",
         "gwaio_anti_hover",
+        "gwaio_anti_naval",
         "gwaio_anti_vehicles",
         "gwaio_combat_titans",
         "gwaio_cooldown_air",