+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -19,7 +19,8 @@ define([
     "gwaio_anti_vehicles",
     "gwaio_anti_bots",
     "gwaio_anti_hover",
-    "gwaio_anti_sea"
+    "gwaio_anti_sea",
+    "gwaio_anti_air"
   );
 
   return {
@@ -456,6 +457,7 @@ define([
       { id: "gwaio_anti_bots" },
       { id: "gwaio_anti_hover" },
       { id: "gwaio_anti_sea" },
+      { id: "gwaio_anti_air" },
     ],
   };
 });