+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/gwo_panel.js
@@ -25,7 +25,7 @@ function gwoWarInfoPanel() {
         // War Information
         const galaxy = game.galaxy();
         const originSystem = galaxy.stars()[galaxy.origin()].system();
-        model.gwoVersion = ko.observable("5.71.0");
+        model.gwoVersion = ko.observable("5.72.0");
         model.gwoSettings = originSystem.gwaio;
 
         if (model.gwoSettings) {