+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - New op multiplyOrAdd which behaves like the old multiply op
+- New tech cards
+  - Anti-Tank Ammo Tech
 
 ### Fixed