+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Anti-Bot Ammo Tech
   - Anti-Hover Ammo Tech
   - Anti-Ship Ammo Tech
+  - Anti-Air Ammo Tech
 
 ### Fixed